Move logic to service and rename event to analytics
