+++ front/src/App.tsx
@@ -2,7 +2,7 @@ import { Navigate, Route, Routes } from 'react-router-dom';
 
 import { RequireAuth } from '@/auth/components/RequireAuth';
 import { RequireNotAuth } from '@/auth/components/RequireNotAuth';
-import { useTrackPageView } from '@/events/hooks/useTrackPageView';
+import { useTrackPageView } from '@/analytics/hooks/useTrackPageView';
 import { useGoToHotkeys } from '@/hotkeys/hooks/useGoToHotkeys';
 import { DefaultLayout } from '@/ui/layout/DefaultLayout';
 import { Index } from '~/pages/auth/Index';
