+++ front/src/modules/analytics/hooks/useEventTracker.ts
