+++ front/src/modules/analytics/hooks/useIsTelemetryEnabled.ts
