+++ front/src/modules/analytics/hooks/useTrackEvent.ts
