+++ front/src/modules/analytics/hooks/useTrackPageView.ts
