+++ front/src/modules/analytics/services/index.ts
