+++ server/src/app.module.ts
@@ -10,7 +10,7 @@ import { GraphQLError } from 'graphql';
 import { PrismaModule } from './database/prisma.module';
 import { HealthModule } from './health/health.module';
 import { AbilityModule } from './ability/ability.module';
-import { EventModule } from './core/event/event.module';
+import { EventModule } from './core/analytics/event.module';
 import GraphQLJSON from 'graphql-type-json';
 
 @Module({
