+++ server/src/core/analytics/dto/create-event.input.ts
