+++ server/src/core/analytics/event.entity.ts
