+++ server/src/core/analytics/event.module.ts
