+++ server/src/core/analytics/event.service.spec.ts
